requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
huggingface_hub
langchain-community
InstructorEmbedding
sentence_transformers